Rename ACARS Hub python files
