+++ rootfs/webapp/acarshub_helpers.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 # Copyright (C) 2022 Frederick Clausen II
-# This file is part of acarshub <https://github.com/fredclausen/docker-acarshub>.
+# File is part of acarshub <https://github.com/fredclausen/docker-acarshub>.
 #
 # acarshub is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,189 +16,457 @@
 # You should have received a copy of the GNU General Public License
 # along with acarshub.  If not, see <http://www.gnu.org/licenses/>.
 
-import logging
 import os
-import requests
+import subprocess
+import acarshub_database
+import time
 
-# debug levels
+start_time = time.time()
 
-DEBUG_LOGGING = False
-EXTREME_LOGGING = False
-QUIET_LOGS = False
-SPAM = False
-ENABLE_ACARS = False
-ENABLE_VDLM = False
-DB_SAVEALL = False
-ACARSHUB_DB = ""
-IATA_OVERRIDE = ""
-DB_BACKUP = ""
-ALERT_STAT_TERMS = []
-ENABLE_ADSB = False
-ADSB_URL = "http://tar1090/data/aircraft.json"
-ADSB_LAT = 0
-ADSB_LON = 0
-ADSB_BYPASS_URL = False
-ACARS_WEB_PORT = 8888  # default port for nginx proxying. SPAM will change this to 80 for running outside of docker
-LIVE_DATA_SOURCE = "127.0.0.1"  # This is to switch from localhost for ACARS/VDLM to connecting to a remote data source
-ACARSHUB_VERSION = "0"
-ACARSHUB_BUILD = "0"
-CURRENT_ACARS_HUB_VERSION = "0"
-CURRENT_ACARS_HUB_BUILD = "0"
-IS_UPDATE_AVAILABLE = False
-ENABLE_RANGE_RINGS = True
-FEED = False
-ARCH = "unknown"
+decoders = dict()
+servers = dict()
+receivers = dict()
+feeders = dict()
+stats = dict()
+external_formats = dict()
+system_error = False
 
-logger = logging.getLogger("werkzeug")
+# Set the URL used for the web front end to show flight tracking.
+# Not doing too much sanitizing of the URL to ensure it's formatted right
+# But we'll at least check and see if the user put a trailing slash or not
 
 
-def log(msg, source):
-    logger.error(f"[{source}]: {msg}")
+if os.getenv("TAR1090_URL", default=False):
+    if os.getenv("TAR1090_URL").endswith("/"):
+        ADSB_URL = os.getenv("TAR1090_URL") + "?icao="
+    else:
+        ADSB_URL = os.getenv("TAR1090_URL") + "/?icao="
+else:
+    ADSB_URL = "https://globe.adsbexchange.com/?icao="
 
 
-if os.getenv("FEED", default=False):
-    FEED = True
+def libacars_formatted(libacars=None):
+    import pprint
 
-if os.getenv("DEBUG_LOGGING", default=False):
-    DEBUG_LOGGING = True
-if os.getenv("EXTREME_LOGGING", default=False):
-    EXTREME_LOGGING = True
-if os.getenv("QUIET_LOGS", default=False):
-    QUIET_LOGS = True
+    html_output = "<p>Decoded:</p>"
+    html_output += "<p>"
+    html_output += "<pre>{libacars}</pre>".format(
+        libacars=pprint.pformat(libacars, indent=2)
+    )
+    html_output += "</p>"
 
-# Application states
+    return html_output
+
+
+# Function to prepare a message for the front end
+# We do any pre-processing/updating of the keys that can't be done on the front end
+
+
+def update_keys(json_message):
+    # Santiztize the message of any empty/None vales
+    # This won't occur for live messages but if the message originates from a DB query
+    # It will return all keys, even ones where the original message didn't have a value
+    stale_keys = []
+    for key in json_message:
+        if json_message[key] is None or json_message[key] == "":
+            stale_keys.append(key)
+
+    for key in stale_keys:
+        del json_message[key]
+
+    # Now we process individual keys, if that key is present
+
+    # database tablename for the message text doesn't match up with typescript-decoder (needs it to be text)
+    # so we rewrite the key
+    if "msg_text" in json_message and json_message["msg_text"] is not None:
+        json_message["text"] = json_message["msg_text"]
+        del json_message["msg_text"]
+
+    if "time" in json_message and json_message["time"] is not None:
+        json_message["timestamp"] = json_message["time"]
+        del json_message["time"]
+
+    if "libacars" in json_message and json_message["libacars"] is not None:
+        json_message["libacars"] = libacars_formatted(json_message["libacars"])
+
+    if "icao" in json_message and json_message["icao"] is not None:
+        json_message["icao_hex"] = format(int(json_message["icao"]), "X")
+
+    if (
+        "flight" in json_message
+        and json_message["flight"] is not None
+        and "icao_hex" in json_message
+    ):
+        json_message["flight"], json_message["icao_flight"] = flight_finder(
+            callsign=json_message["flight"], hex_code=json_message["icao_hex"]
+        )
+    elif "flight" in json_message and json_message["flight"] is not None:
+        json_message["flight"], json_message["icao_flight"] = flight_finder(
+            callsign=json_message["flight"], url=False
+        )
+    elif "icao_hex" in json_message:
+        json_message["icao_url"] = flight_finder(hex_code=json_message["icao_hex"])
+
+    if "toaddr" in json_message and json_message["toaddr"] is not None:
+        json_message["toaddr_hex"] = format(int(json_message["toaddr"]), "X")
+
+        toaddr_icao, toaddr_name = acarshub_database.lookup_groundstation(
+            json_message["toaddr_hex"]
+        )
+
+        if toaddr_icao is not None:
+            json_message["toaddr_decoded"] = f"{toaddr_name} ({toaddr_icao})"
+
+    if "fromaddr" in json_message and json_message["fromaddr"] is not None:
+        json_message["fromaddr_hex"] = format(int(json_message["fromaddr"]), "X")
+
+        fromaddr_icao, fromaddr_name = acarshub_database.lookup_groundstation(
+            json_message["fromaddr_hex"]
+        )
+
+        if fromaddr_icao is not None:
+            json_message["fromaddr_decoded"] = f"{fromaddr_name} ({fromaddr_icao})"
+
+    if "label" in json_message and json_message["label"] is not None:
+        label_type = acarshub_database.lookup_label(json_message["label"])
+
+        if label_type is not None:
+            json_message["label_type"] = label_type
+        else:
+            json_message["label_type"] = "Unknown Message Label"
+
+    return json_message
+
+
+def flight_finder(callsign=None, hex_code=None, url=True):
+    global ADSB_URL
+
+    # If there is only a hex code, we'll return just the ADSB url
+    # Front end will format correctly.
+
+    if callsign is None and hex_code is not None:
+        return f"{ADSB_URL}{hex_code}"
+
+    if callsign is not None:
+        # Check the ICAO DB to see if we know what it is
+        # The ICAO DB will return the ICAO code back if it didn't find anything
+
+        icao, airline = acarshub_database.find_airline_code_from_iata(callsign[:2])
+        flight_number = callsign[2:]
+        flight = icao + flight_number
+        tooltip_text = ""
+
+        if icao != callsign[:2]:
+            html = f"<strong>{flight}/{callsign}</strong> "
+            tooltip_text = (
+                f"<p>The aircraft's callsign.</p>{airline} Flight {flight_number}"
+            )
+        else:
+            html = f"<strong>{flight}</strong> "
+            tooltip_text = f"<p>The aircraft's callsign was not found in the database for decoding.</p>{flight}"
+
+        # If the iata and icao variables are not equal, airline was found in the database and we'll add in the tool-tip for the decoded airline
+        # Otherwise, no tool-tip, no FA link, and use the IATA code for display
+        if url:
+            return (
+                f'<span class="flight-tooltip" data-jbox-content="{tooltip_text}">Flight: <strong><a href="{ADSB_URL}{hex_code}" target="_blank">{html}</a></strong></span>',
+                flight,
+            )
+        else:
+            return (
+                f'<span class="flight-tooltip" data-jbox-content="{tooltip_text}">Flight: {html}</span>',
+                flight,
+            )
+    else:  # We should never run in to this condition, I don't think, but we'll add a case for it
+        return ("Flight: Error", None)
+
+
+def handle_message(message=None):
+    if message is not None:
+        total_results = 0
+        serialized_json = []
+        search_term = ""
+        # If the user has cleared the search bar, we'll execute the else statement
+        # And the browsers clears the data
+        # otherwise, run the search and format the results
+
+        if "show_all" in message or message["search_term"] != "":
+            # Decide if the user is executing a new search or is clicking on a page of results
+            # search.js appends the "results_after" as the result page index user is requested
+            # Otherwise, we're at the first page
+
+            if "search_term" in message:
+                search_term = message["search_term"]
+
+                if "results_after" in message:
+                    # ask the database for the results at the user requested page
+                    results, total_results = acarshub_database.database_search(
+                        message["search_term"], message["results_after"]
+                    )
+                else:
+                    results, total_results = acarshub_database.database_search(
+                        message["search_term"]
+                    )
+            elif "show_all" in message:
+                if "results_after" in message:
+                    results, total_results = acarshub_database.show_all(
+                        message["results_after"]
+                    )
+                else:
+                    results, total_results = acarshub_database.show_all()
+
+            # the db returns two values
+            # index zero is the query results in json
+            # the other is the count of total results
+
+            if results is not None:
+                # Loop through the results and format html
+                for result in results:
+                    serialized_json.append(update_keys(result))
+
+            return (total_results, serialized_json, search_term)
+    else:
+        return (None, None, None)
+
+
+def service_check():
+    import re
+
+    global decoders
+    global servers
+    global receivers
+    global system_error
+    global feeders
+    global stats
+    global start_time
+    global external_formats
+
+    if os.getenv("SPAM", default=False):
+        healthcheck = subprocess.Popen(
+            ["../../tools/healthtest.sh"],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+    else:
+        healthcheck = subprocess.Popen(
+            ["/scripts/healthcheck.sh"], stdout=subprocess.PIPE, stderr=subprocess.PIPE
+        )
+    stdout, stderr = healthcheck.communicate()
+    healthstatus = stdout.decode()
+
+    decoders = dict()
+    servers = dict()
+    receivers = dict()
+    feeders = dict()
+    stats = dict()
+    external_formats = dict()
+    system_error = False
+
+    for line in healthstatus.split("\n"):
+        try:
+            match = re.search("(?:acarsdec|dumpvdl2)-.+ =", line)
+            if match:
+                if match.group(0).strip(" =") not in decoders:
+                    decoders[match.group(0).strip(" =")] = dict()
+                    continue
+            else:
+                for decoder in decoders:
+                    if line.find(decoder) != -1:
+                        if line.find(f"Decoder {decoder}") and line.endswith(
+                            "UNHEALTHY"
+                        ):
+                            decoders[decoder]["Status"] = "Bad"
+                            system_error = True
+                        elif line.find(f"Decoder {decoder}") == 0 and line.endswith(
+                            "HEALTHY"
+                        ):
+                            decoders[decoder]["Status"] = "Ok"
+                        elif line.find(f"Decoder {decoder}") == 0:
+                            system_error = True
+                            decoders[decoder]["Status"] = "Unknown"
+
+                        continue
+
+            match = re.search("^(?:acars|vdlm2)_server", line)
+
+            if match:
+                if match.group(0) not in servers:
+                    servers[match.group(0)] = dict()
+
+                if line.find("listening") != -1 and line.endswith("UNHEALTHY"):
+                    servers[match.group(0)]["Status"] = "Bad"
+                    system_error = True
+                elif line.find("listening") != -1 and line.endswith("HEALTHY"):
+                    servers[match.group(0)]["Status"] = "Ok"
+                elif line.find("listening") != -1:
+                    system_error = True
+                    servers[match.group(0)]["Status"] = "Unknown"
+                elif line.find("python") != -1 and line.endswith("UNHEALTHY"):
+                    system_error = True
+                    servers[match.group(0)]["Web"] = "Bad"
+                elif line.find("python") != -1 and line.endswith("HEALTHY"):
+                    servers[match.group(0)]["Web"] = "Ok"
+                elif line.find("python") != -1:
+                    system_error = True
+                    servers[match.group(0)]["Web"] = "Unknown"
+
+                continue
+
+            match = re.search("\\d+\\s+(?:ACARS|VDLM2) messages", line)
+
+            if match:
+                if line.find("ACARS") != -1 and "ACARS" not in receivers:
+                    receivers["ACARS"] = dict()
+                    receivers["ACARS"]["Count"] = line.split(" ")[0]
+                    if line.endswith("UNHEALTHY"):
+                        if time.time() - start_time > 300.0:
+                            system_error = True
+                            receivers["ACARS"]["Status"] = "Bad"
+                        else:
+                            receivers["ACARS"]["Status"] = "Waiting for first message"
+                    elif line.endswith("HEALTHY"):
+                        receivers["ACARS"]["Status"] = "Ok"
+                    else:
+                        system_error = True
+                        receivers["ACARS"]["Status"] = "Unknown"
+                if line.find("VDLM2") != -1 and "VDLM2" not in receivers:
+                    receivers["VDLM2"] = dict()
+                    receivers["VDLM2"]["Count"] = line.split(" ")[0]
+                    if line.endswith("UNHEALTHY"):
+                        if time.time() - start_time > 300.0:
+                            system_error = True
+                            receivers["VDLM2"]["Status"] = "Bad"
+                        else:
+                            receivers["VDLM2"]["Status"] = "Waiting for first message"
+                    elif line.endswith("HEALTHY"):
+                        receivers["VDLM2"]["Status"] = "Ok"
+                    else:
+                        system_error = True
+                        receivers["VDLM2"]["Status"] = "Unknown"
+
+                continue
+
+            match = re.search("^(?:acars|vdlm2)_feeder", line)
+
+            if match:
+                if match.group(0) not in servers:
+                    feeders[match.group(0)] = dict()
+
+                if line.endswith("UNHEALTHY"):
+                    feeders[match.group(0)]["Status"] = "Bad"
+                    system_error = True
+                elif line.endswith("HEALTHY"):
+                    feeders[match.group(0)]["Status"] = "Ok"
+                else:
+                    system_error = True
+                    feeders[match.group(0)]["Status"] = "Unknown"
+
+                continue
+
+            match = re.search("^(acars|vdlm2)_stats", line)
+
+            if match:
+                if os.getenv("DEBUG_LOGGING", default=False):
+                    print(line)
+                if match.group(0) not in stats:
+                    stats[match.group(0)] = dict()
+
+                if line.endswith("UNHEALTHY"):
+                    system_error = True
+                    stats[match.group(0)]["Status"] = "Bad"
+                elif line.endswith("HEALTHY"):
+                    stats[match.group(0)]["Status"] = "Ok"
+                else:
+                    system_error = True
+                    stats[match.group(0)]["Status"] = "Unknown"
+
+            match = re.search("^planeplotter", line)
+
+            if match:
+                if line.find("vdl2") != -1:
+                    pp_decoder = "VDLM2"
+                else:
+                    pp_decoder = "ACARS"
+
+                if pp_decoder not in external_formats:
+                    external_formats[pp_decoder] = []
+
+                if line.endswith("UNHEALTHY"):
+                    system_error = True
+                    external_formats[pp_decoder].append(
+                        {"type": "planeplotter", "Status": "Bad"}
+                    )
+                elif line.endswith("HEALTHY"):
+                    external_formats[pp_decoder].append(
+                        {"type": "planeplotter", "Status": "Ok"}
+                    )
+                else:
+                    system_error = True
+                    external_formats[pp_decoder].append(
+                        {"type": "planeplotter", "Status": "Unknown"}
+                    )
+
+            match = re.search("dumpvdl2 and planeplotter", line)
+
+            if match:
+                if line.find("vdl2") != -1:
+                    pp_decoder = "VDLM2"
+                else:
+                    pp_decoder = "ACARS"
+
+                if pp_decoder not in external_formats:
+                    external_formats[pp_decoder] = []
+
+                if line.endswith("UNHEALTHY"):
+                    system_error = True
+                    external_formats[pp_decoder].append(
+                        {"type": "dumpvdl2 to planeplotter", "Status": "Bad"}
+                    )
+                elif line.endswith("HEALTHY"):
+                    external_formats[pp_decoder].append(
+                        {"type": "dumpvdl2 to planeplotter", "Status": "Ok"}
+                    )
+                else:
+                    system_error = True
+                    external_formats[pp_decoder].append(
+                        {"type": "dumpvdl2 to planeplotter", "Status": "Unknown"}
+                    )
+
+        except Exception as e:
+            print(f"[service-check] Error: {line}\n{e}")
+
+    if os.getenv("SPAM", default=False):
+        print(decoders)
+        print(servers)
+        print(receivers)
+        print(feeders)
+        print(stats)
+        print(external_formats)
+
 
 if os.getenv("SPAM", default=False):
-    SPAM = True
-    ACARS_WEB_PORT = 80
-if os.getenv("LIVE_DATA_SOURCE", default=False):
-    LIVE_DATA_SOURCE = os.getenv("LIVE_DATA_SOURCE")
-if os.getenv("ENABLE_ACARS", default=False):
-    ENABLE_ACARS = True
-if os.getenv("ENABLE_VDLM", default=False):
-    ENABLE_VDLM = True
-
-if os.getenv("DB_SAVEALL", default=False):
-    DB_SAVEALL = True
-
-# Application Settings
-
-if os.getenv("ACARSHUB_DB", default=False):
-    ACARSHUB_DB = os.getenv("ACARSHUB_DB", default=False)
-else:
-    ACARSHUB_DB = "sqlite:////run/acars/messages.db"
-
-if os.getenv("DB_BACKUP", default=False):
-    DB_BACKUP = os.getenv("DB_BACKUP")
-
-if os.getenv("IATA_OVERRIDE", default=False):
-    IATA_OVERRIDE = os.getenv("IATA_OVERRIDE")
-
-if os.getenv("ALERT_STAT_TERMS", default=False):
-    ALERT_STAT_TERMS = os.getenv("ALERT_STAT_TERMS").split(",")
-else:
-    ALERT_STAT_TERMS = [
-        "cop",
-        "police",
-        "authorities",
-        "chop",
-        "turbulence",
-        "turb",
-        "fault",
-        "divert",
-        "mask",
-        "csr",
-        "agent",
-        "medical",
-        "security",
-        "mayday",
-        "emergency",
-        "pan",
-        "red coat",
-    ]
-
-if os.getenv("ENABLE_ADSB", default=False) == "true":
-    ENABLE_ADSB = True
-    if os.getenv("ADSB_URL", default=False):
-        ADSB_URL = os.getenv("ADSB_URL", default=False)
-
-        if not ADSB_URL.startswith("http") and not ADSB_URL.endswith("aircraft.json"):
-            ENABLE_ADSB = False
-            log(
-                f"ADSB URL ({ADSB_URL}) appears to be malformed. Disabling ADSB", "init"
-            )
-    if os.getenv("ADSB_LON", default=False):
-        ADSB_LON = float(os.getenv("ADSB_LON"))
-    if os.getenv("ADSB_LAT", default=False):
-        ADSB_LAT = float(os.getenv("ADSB_LAT"))
-    if os.getenv("DISABLE_RANGE_RINGS", default=False):
-        ENABLE_RANGE_RINGS = False
-
-if os.getenv("ADSB_BYPASS_URL", default=False):
-    ADSB_BYPASS_URL = True
+    service_check()
 
 
-if SPAM:
-    version_path = "../../VERSION"
-else:
-    version_path = "/acarshub-version"
-with open(version_path, "r") as f:
-    lines = f.read()
-    ACARSHUB_VERSION = lines.split("\n")[0].split(" ")[0].replace("v", "")
-    CURRENT_ACARS_HUB_VERSION = ACARSHUB_VERSION
-    ACARSHUB_BUILD = lines.split("\n")[0].split(" ")[2].replace("v", "")
-    CURRENT_ACARS_HUB_BUILD = ACARSHUB_BUILD
+def get_service_status():
+    global decoders
+    global servers
+    global receivers
+    global system_error
+    global feeders
+    global stats
+    global external_formats
 
-if not SPAM and os.path.exists("/arch"):
-    with open("/arch", "r") as f:
-        lines = f.read()
-        ARCH = lines.split("\n")[0]
-
-
-def check_github_version():
-    global CURRENT_ACARS_HUB_VERSION
-    global ACARSHUB_VERSION
-    global ACARSHUB_BUILD
-    global CURRENT_ACARS_HUB_BUILD
-    global IS_UPDATE_AVAILABLE
-    # FIXME: This is a hack to get around the fact that the version file is not updated on the build server
-    if not SPAM:
-        r = requests.get(
-            "https://raw.githubusercontent.com/fredclausen/docker-acarshub/main/version"
-        )
-        CURRENT_ACARS_HUB_VERSION = r.text.split("\n")[0].split(" ")[0].replace("v", "")
-        CURRENT_ACARS_HUB_BUILD = r.text.split("\n")[0].split(" ")[2].replace("v", "")
-
-        if (
-            CURRENT_ACARS_HUB_VERSION != ACARSHUB_VERSION
-            and ACARSHUB_VERSION < CURRENT_ACARS_HUB_VERSION
-        ) or (
-            CURRENT_ACARS_HUB_BUILD != ACARSHUB_BUILD
-            and ACARSHUB_BUILD < CURRENT_ACARS_HUB_BUILD
-        ):
-            log("Update found", "version-checker")
-            IS_UPDATE_AVAILABLE = True
-        else:
-            if not QUIET_LOGS:
-                log("No update found", "version-checker")
-            IS_UPDATE_AVAILABLE = False
-
-
-def get_version():
-    global CURRENT_ACARS_HUB_VERSION
-    global CURRENT_ACARS_HUB_BUILD
-    global ACARSHUB_VERSION
-    global ACARSHUB_BUILD
-    global IS_UPDATE_AVAILABLE
     return {
-        "github_version": "v"
-        + CURRENT_ACARS_HUB_VERSION
-        + " Build "
-        + CURRENT_ACARS_HUB_BUILD,
-        "container_version": "v" + ACARSHUB_VERSION + " Build " + ACARSHUB_BUILD,
-        "is_outdated": IS_UPDATE_AVAILABLE,
+        "decoders": decoders,
+        "servers": servers,
+        "global": receivers,
+        "feeders": feeders,
+        "error_state": system_error,
+        "stats": stats,
+        "external_formats": external_formats,
     }
 
 
-def acars_traceback(e, source):
-    logger.exception(f"[{source}]: {e}")
+def getAlerts():
+    return acarshub_database.get_alert_counts()
