+++ rootfs/webapp/acarshub_rrd_database.py
@@ -17,7 +17,7 @@
 # along with acarshub.  If not, see <http://www.gnu.org/licenses/>.
 
 import rrdtool
-import acarshub_helpers
+import acarshub_configuration
 import os
 
 
@@ -26,18 +26,18 @@ def update_db(vdlm=0, acars=0, error=0):
 
     try:
         rrdtool.update("/run/acars/acarshub.rrd", f"N:{acars}:{vdlm}:{total}:{error}")
-        if acarshub_helpers.DEBUG_LOGGING:
-            acarshub_helpers.log(
+        if acarshub_configuration.DEBUG_LOGGING:
+            acarshub_configuration.log(
                 f"rrdtool.update: N:{acars}:{vdlm}:{total}:{error}", "rrdtool"
             )
     except Exception as e:
-        acarshub_helpers.acars_traceback(e, "rrdtool")
+        acarshub_configuration.acars_traceback(e, "rrdtool")
 
 
 def create_db():
     try:
         if not os.path.exists("/run/acars/acarshub.rrd"):
-            acarshub_helpers.log("creating the RRD Database", "rrdtool")
+            acarshub_configuration.log("creating the RRD Database", "rrdtool")
             rrdtool.create(
                 "/run/acars/acarshub.rrd",
                 "--start",
@@ -54,7 +54,7 @@ def create_db():
                 "RRA:AVERAGE:0.5:360:4380",  # 3 year at 6 hour reso
             )
     except Exception as e:
-        acarshub_helpers.acars_traceback(e, "rrdtool")
+        acarshub_configuration.acars_traceback(e, "rrdtool")
     else:
-        if not acarshub_helpers.QUIET_LOGS:
-            acarshub_helpers.log("Database found", "rrdtool")
+        if not acarshub_configuration.QUIET_LOGS:
+            acarshub_configuration.log("Database found", "rrdtool")
